Temporarily remove SIGKILLing after a timeout
At present, when killing a container we SIGTERM, and then if the
container is still running after 10s we SIGKILL. This is potentially
useful (but might need per-job configuring.)

Sadly right now we don't keep enough state for this to be safe. It's
possible that the SIGKILL would happen to the newly-recreated
container. We probably want to track unique IDs for each task instance
and/or keep a state for each task.